refactor(runner): split idle pool lifecycle modules
